refactor - MatchBonus 파일 분리
